chore: rename multibanco apm (#1889)
