+++ __sdk__.js
@@ -28,7 +28,7 @@ const SMART_FUNDING_SOURCES = [
     FUNDING.TRUSTLY,
     FUNDING.OXXO,
     FUNDING.BOLETO,
-    FUNDING.MULTIBLANCO,
+    FUNDING.MULTIBANCO,
     FUNDING.MAXIMA,
     FUNDING.MERCADOPAGO,
     FUNDING.CARD
