+++ test/globals.js
@@ -77,7 +77,7 @@ const fundingEligibility = {
     boleto: {
         eligible: false
     },
-    multiblanco: {
+    multibanco: {
         eligible: false
     },
     maxima: {
