+++ test/integration/globals.js
@@ -74,7 +74,7 @@ window.__TEST_FUNDING_ELIGIBILITY__ = {
     boleto: {
         eligible: false
     },
-    multiblanco: {
+    multibanco: {
         eligible: false
     },
     maxima: {
